Add NavBar structure and feature, Add Home.html, Add Quote.html, move Calculator from components/ to page/
