+++ src/App.js
@@ -1,6 +1,10 @@
 import React from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Navbar from './components/Navbar';
 import './App.css';
-import Calculator from './components/Calculator';
+import Home from './pages/Home'
+import Calculator from './pages/Calculator';
+import Quote from './pages/Quote'
 
 class App extends React.Component {
   constructor(props) {
@@ -10,9 +14,14 @@ class App extends React.Component {
 
   render() {
     return (
-      <div className="App">
-        <Calculator />
-      </div>
+      <BrowserRouter className="App">
+          <Navbar />
+        <Routes>
+          <Route path='/' element={<Home />} />
+          <Route path='/quote' element={<Quote />} />
+          <Route path='/calculator' element={<Calculator />} />
+        </Routes>
+      </BrowserRouter>
     );
   }
 }
