+++ src/pages/Home.js
@@ -1,0 +1,10 @@
+
+const Home = () => {
+  return (
+    <main>
+      <h2>This is the home page</h2>
+    </main>
+  )
+}
+
+export default Home
