+++ src/pages/Quote.js
@@ -1,0 +1,10 @@
+
+const Quote = () => {
+  return (
+    <main>
+      
+    </main>
+  )
+}
+
+export default Quote;
